bugfix:优化修改shard分布的采集任务

+++ arius-admin/arius-admin-persistence/src/main/java/com/didichuxing/datachannel/arius/admin/persistence/es/index/dao/stats/BaseAriusStatsESDAO.java
@@ -762,7 +762,7 @@ public class BaseAriusStatsESDAO extends BaseESDAO {
             }
             //get indexCount
             if (INDEX_COUNT.equals(key)&&null != esAggr && null != esAggr.getUnusedMap().get(VALUE)) {
-                metricsContent.setIndexCount(Double.parseDouble(esAggr.getUnusedMap().get(VALUE).toString()));
+                metricsContent.setIndexCount(Long.parseLong(esAggr.getUnusedMap().get(VALUE).toString()));
             }
 
             metricsContentCells.add(metricsContentCell);
